% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedbar.R
\name{stackedbar}
\alias{stackedbar}
\title{Calculates the cumulative mean values from a data frame and creates a stacked bar plot based upon those values allowing for error bars to added.}
\usage{
stackedbar(
  df,
  columns,
  x,
  standard_error = "sem",
  graph = "TRUE",
  y_axis_title = "y",
  fill_name = "fill"
)
}
\arguments{
\item{df}{A data frame.}

\item{columns}{A character vector. A list of column numbers or names containing the data.}

\item{x}{A number or string. The x axis variable name or column number.}

\item{standard_error}{The measurement of variation, either "sem" or "sd".}

\item{graph}{A string. Determines if the output is a graph or the data frame. Either "TRUE" or "FALSE".}

\item{y_axis_title}{A string. The name of the y axis. Defaults to "y".}

\item{fill_name}{A string. The name of the variable to colour by. Defaults to "fill".}
}
\value{
A stacked bar graph or data frame.
}
\description{
Calculates the cumulative mean values from a data frame and creates a stacked bar plot based upon those values allowing for error bars to added.
}
\examples{
stackedbar(iris, c(2,3), 5, standard_error = "sem", graph = "TRUE", y_axis_title = "Measurement", fill_name = "Attributes")

stackedbar(iris,c("Sepal.Width", "Sepal.Length"), "Species", standard_error = "sd", graph = "FALSE")
}
