% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartenderr.R
\name{stackedbar}
\alias{stackedbar}
\title{Calculates the cumulative mean values from a data frame and creates a stacked bar plot based upon those values allowing for error bars to added.}
\usage{
stackedbar(df, x, y, fill_by, standard_error = "sem", graph = "TRUE")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A character vector. The x axis variable name or column number.}

\item{y}{A numeric vector. The y axis variable name or column number.}

\item{fill_by}{A character vector. The name or column number of the stacking variable.}

\item{standard_error}{The measurement of variation, either "sem" or "sd".}

\item{graph}{Determines if the output is a graph or the data frame. Either "TRUE" or "FALSE".}
}
\value{
A stacked bar graph or data frame.
}
\description{
Calculates the cumulative mean values from a data frame and creates a stacked bar plot based upon those values allowing for error bars to added.
}
\examples{
stackedbar(iris, 5, 2, 5, standard_error = "sem", graph = "TRUE")

stackedbar(iris,"Species", "Sepal.Width", "Species", standard_error = "sd", graph = "FALSE")
}
